# specs/001-fraud-review-platform/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Fraud Review API
  version: 0.1.0
  description: Contract for the local fraud scoring and analyst-review MVP.
servers:
  - url: http://localhost:8000/api/v1
paths:
...
  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      parameters:
        - $ref: "#/components/parameters/Start"
        - $ref: "#/components/parameters/End"
      responses:
        "200":
          description: Reconciled metrics for an inclusive-start/exclusive-end UTC range
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
  /models/active:
    get:
      operationId: getActiveModel
      responses:
        "200":
          description: Active model identity and evaluation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelSummary"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Active model has no complete labeled evaluation
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
components:
  parameters:
...
    Start:
      name: start
      in: query
      required: true
      description: Inclusive UTC instant
      schema:
        type: string
        format: date-time
    End:
      name: end
      in: query
      required: true
      description: Exclusive UTC instant
      schema:
        type: string
        format: date-time
  responses:
    NotFound:
      description: Resource not found
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
...
    DashboardSummary:
      type: object
      required: [range, totals, riskBands, reviewOutcomes, series]
      properties:
        range:
          type: object
          required: [start, end]
          properties:
            start: {type: string, format: date-time}
            end: {type: string, format: date-time}
        totals:
          type: object
          required: [transactions, alerts, amountAtRisk]
          properties:
            transactions: {type: integer, minimum: 0}
            alerts: {type: integer, minimum: 0}
            amountAtRisk: {type: string}
        riskBands: {type: object, additionalProperties: {type: integer, minimum: 0}}
        reviewOutcomes: {type: object, additionalProperties: {type: integer, minimum: 0}}
        series:
          type: array
          items:
            type: object
            required: [bucket, transactions, alerts]
            properties:
              bucket: {type: string, format: date-time}
              transactions: {type: integer, minimum: 0}
              alerts: {type: integer, minimum: 0}
    ModelSummary:
      type: object
      required: [version, featureVersion, datasetId, threshold, metrics, activatedAt]
      properties:
        version: {type: string}
        featureVersion: {type: string}
        datasetId: {type: string}
        threshold: {type: number, minimum: 0, maximum: 1}
        metrics:
          type: object
          required: [precision, recall, prAuc, truePositive, falsePositive, trueNegative, falseNegative, alertVolume]
          properties:
            precision: {type: number, minimum: 0, maximum: 1}
            recall: {type: number, minimum: 0, maximum: 1}
            prAuc: {type: number, minimum: 0, maximum: 1}
            truePositive: {type: integer, minimum: 0}
            falsePositive: {type: integer, minimum: 0}
            trueNegative: {type: integer, minimum: 0}
            falseNegative: {type: integer, minimum: 0}
            alertVolume: {type: integer, minimum: 0}
        activatedAt: {type: string, format: date-time}
    Problem:
      type: object
      required: [type, title, status, detail]
      properties:
        type: {type: string, format: uri-reference}
        title: {type: string}
        status: {type: integer}
        detail: {type: string}
        errors:
          type: array
          items:
            type: object
            required: [field, message]
            properties:
              field: {type: string}
              message: {type: string}